\name{dgp}
\alias{dgp}
\title{
Sample data generating process
}
\description{
Sample data generating process
}
\usage{
dgp(n, sd = 1)
}
\arguments{
  \item{n}{
sample size
}
  \item{sd}{
standard deviation
}
}
\value{
data.frame
}
\seealso{
\code{\link{hgam}}, \code{\link{rnorm}}, \code{\link{sin}}
}
\keyword{data}