\name{hgam-package}
\alias{hgam-package}
\docType{package}
\title{
Fitting high-dimensional generalized additive models
}
\description{
\code{hgam} is used to fit high-dimensional generalized additive models.
}
\details{
\tabular{ll}{
Package: \tab hgam\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2010-02-09\cr
License: \tab GPL-2\cr
}
The package implements high-dimensional additive models
as introduced by Meier et al. (2009).
}
\references{

    Meier, L., van de Geer, S. and Buehlmann, P. (2009). 
    High-dimensional additive modeling. \emph{Annals of Statistics} 
    \bold{37}, 3779--3821.
}
\author{
The students of the `Advanced R Programming Course' (winter term 2009/2010)
at LMU Muenchen:
Hannah Frick, Ivan Kondofersky, Oliver S. Kuehnle,
Christian Lindenlaub, Georg Pfundstein, 
Matthias Speidel, Martin Spindler, Ariane Straub,
Florian Wickler, Katharina Zink
\cr
under the supervision of Manuel Eugster 
<Manuel.Eugster@stat.uni-muenchen.de> (package maintainer) and
Torsten Hothorn.
}
\keyword{ package }
